Deleting a friend and show MessageBox before delete to ask confirmation

+++ FriendStorage/FriendStorage.UI/ViewModel/FriendEditViewModel.cs
@@ -6,6 +6,8 @@ using System.Windows.Input;
 using FriendStorage.UI.Wrapper;
 using Prism.Events;
 using FriendStorage.UI.Events;
+using System.Windows;
+using FriendStorage.UI.Dialogs;
 
 namespace FriendStorage.UI.ViewModel
 {
@@ -19,12 +21,15 @@ namespace FriendStorage.UI.ViewModel
         private IFriendDataProvider _dataProvider;
         private FriendWrapper _friend;
         private IEventAggregator _eventAggregator;
+        private IMessageDialogService _messageDialogService;
 
         public FriendEditViewModel(IFriendDataProvider dataProvider,
-          IEventAggregator eventAggregator)
+          IEventAggregator eventAggregator,
+          IMessageDialogService messageDialogService)
         {
             _dataProvider = dataProvider;
             _eventAggregator = eventAggregator;
+            _messageDialogService = messageDialogService;
             SaveCommand = new DelegateCommand(OnSaveExecute, OnSaveCanExecute);
             DeleteCommand = new DelegateCommand(OnDeleteExecute, OnDeleteCanExecute);
         }
@@ -84,8 +89,13 @@ namespace FriendStorage.UI.ViewModel
 
         private void OnDeleteExecute(object obj)
         {
-            _dataProvider.DeleteFriend(Friend.Id);
-            _eventAggregator.GetEvent<FriendDeletedEvent>().Publish(Friend.Id);
+            var result = _messageDialogService.ShowYesNoDialog("Delete Friend",
+              $"Do you really want to delete the friend '{Friend.FirstName} {Friend.LastName}'");
+            if (result == MessageDialogResult.Yes)
+            {
+                _dataProvider.DeleteFriend(Friend.Id);
+                _eventAggregator.GetEvent<FriendDeletedEvent>().Publish(Friend.Id);
+            }
         }
 
         private bool OnDeleteCanExecute(object arg)
